first attempt at fallback to ftp.tug.org for obsolete TeX Live

+++ TLMDatabase.h
@@ -40,6 +40,6 @@
 
 @interface TLMDatabase : NSObject
 
-+ (NSUInteger)yearForMirrorURL:(NSURL *)aURL;
++ (int16_t)yearForMirrorURL:(NSURL *)aURL;
 
 @end
